+++ studio-application-modules/attachment-service/src/main/java/studio/one/application/web/controller/AttachmentMgmtController.java
@@ -185,6 +185,6 @@ public class AttachmentMgmtController {
     }
 
     private AttachmentDto toDto(Attachment attachment) {
-        return AttachmentWebSupport.toDto(attachment, identityServiceProvider, log);
+        return AttachmentWebSupport.toDto(attachment, identityServiceProvider);
     }
 }